requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0

# AI/ML Libraries
ultralytics==8.4.5
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.12.0.88
numpy>=1.24.0
Pillow==11.0.0

# Visualization
matplotlib==3.10.0
seaborn==0.13.0

# PDF Generation
reportlab==4.2.5

# Database
SQLAlchemy==2.0.36
